+++ src/pages/Home/HomePage.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { Layout, Button, Typography, Row, Col, Card } from "antd";
+import { Button, Typography, Row, Col, Card } from "antd";
 import { useNavigate } from "react-router-dom";
-import { ProLayout, PageContainer } from "@ant-design/pro-components";
-import { HomeOutlined, LoginOutlined, UserAddOutlined } from "@ant-design/icons";
+import { PageContainer } from "@ant-design/pro-components";
+import Header from "../../components/Header";
 
 const { Title, Paragraph } = Typography;
 
@@ -50,105 +50,65 @@ function generateProducts(count = 12) {
   });
 }
 
-const menuRoutes = [
-  { path: "/", name: "Trang chủ", icon: <HomeOutlined /> },
-];
-
 const HomePage = () => {
   const navigate = useNavigate();
   const products = React.useMemo(() => generateProducts(12), []);
   return (
-    <div style={{ width: "100vw", marginLeft: "calc(50% - 50vw)", marginRight: "calc(50% - 50vw)", minHeight: "100vh" }}>
-      <ProLayout
-        title="Hệ thống quản lý dịch vụ"
-        route={{ routes: menuRoutes }}
-        logo={() => (
-          <div style={{
-            width: 32,
-            height: 32,
-            borderRadius: 6,
-            background: "#1677ff",
-            color: "#fff",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            fontWeight: 700,
-            fontSize: 14,
-          }}>SW</div>
-        )}
-        layout="top"
-        fixedHeader
-        actionsRender={() => [
-          <Button key="login" onClick={() => navigate('/login')} icon={<LoginOutlined />}>Đăng nhập</Button>,
-          <Button key="register" type="primary" onClick={() => navigate('/register')} icon={<UserAddOutlined />}>Đăng ký</Button>,
-        ]}
-        menuItemRender={(item, dom) => (
-          <a onClick={() => item.path && navigate(item.path)}>{dom}</a>
-        )}
-        footerRender={() => (
-          <div style={{ textAlign: "center", color: "#64748b", padding: "12px 0" }}>
-            ©2024 Created by Your Company
-          </div>
-        )}
-        token={{
-          header: { colorTextMenu: "#0f172a" },
-        }}
-      >
-        <PageContainer breadcrumb={false}>
-          <div className="bg-gradient-to-br from-blue-50 to-white">
-            <div className="text-center w-full px-6 py-16">
-              <Title level={1} className="text-3xl font-bold mb-4">
-                Welcome to Our Service
-              </Title>
-              <Paragraph className="text-lg text-gray-600 mb-8">
-                Please login or register to continue.
-              </Paragraph>
-              <div className="flex justify-center gap-6">
-                <Button
-                  type="primary"
-                  size="large"
-                  onClick={() => navigate("/login")}
-                  className="!min-w-[120px] !rounded-xl !font-medium hover:!scale-105 hover:!shadow-lg transition-all duration-200"
-                >
-                  Login
-                </Button>
-                <Button
-                  type="default"
-                  size="large"
-                  onClick={() => navigate("/register")}
-                  className="!min-w-[120px] !rounded-xl !font-medium hover:!scale-105 hover:!shadow-lg transition-all duration-200"
-                >
-                  Register
-                </Button>
-              </div>
+    <Header title="Hệ thống quản lý dịch vụ">
+      <PageContainer breadcrumb={false}>
+        <div className="bg-gradient-to-br from-blue-50 to-white">
+          <div className="text-center w-full px-6 py-16">
+            <Title level={1} className="text-3xl font-bold mb-4">
+              Welcome to Our Service
+            </Title>
+            <Paragraph className="text-lg text-gray-600 mb-8">
+              Please login or register to continue.
+            </Paragraph>
+            <div className="flex justify-center gap-6">
+              <Button
+                type="primary"
+                size="large"
+                onClick={() => navigate("/login")}
+                className="!min-w-[120px] !rounded-xl !font-medium hover:!scale-105 hover:!shadow-lg transition-all duration-200"
+              >
+                Login
+              </Button>
+              <Button
+                type="default"
+                size="large"
+                onClick={() => navigate("/register")}
+                className="!min-w-[120px] !rounded-xl !font-medium hover:!scale-105 hover:!shadow-lg transition-all duration-200"
+              >
+                Register
+              </Button>
+            </div>
 
-              {/* Products/Services grid */}
-              <div className="w-full mt-14">
-                <Title level={3} className="!mb-4">Sản phẩm / Dịch vụ mẫu</Title>
-                <Paragraph className="text-gray-600 !mb-8">
-                  Danh sách ngẫu nhiên được tạo khi tải trang để minh họa giao diện.
-                </Paragraph>
-                <Row gutter={[16, 16]}>
-                  {products.map((p) => (
-                    <Col key={p.id} xs={24} sm={12} md={8} lg={6}>
-                      <Card hoverable className="!rounded-xl h-full" title={p.name}>
-                        <Paragraph className="!mb-4 text-gray-600">{p.desc}</Paragraph>
-                        <div className="text-xl font-semibold text-blue-600">
-                          {p.price.toLocaleString("vi-VN")}₫
-                        </div>
-                        <div className="mt-4">
-                          <Button type="primary" block className="!rounded-lg">Dùng thử</Button>
-                        </div>
-                      </Card>
-                    </Col>
-                  ))}
-                </Row>
-              </div>
+            {/* Products/Services grid */}
+            <div className="w-full mt-14">
+              <Title level={3} className="!mb-4">Sản phẩm / Dịch vụ mẫu</Title>
+              <Paragraph className="text-gray-600 !mb-8">
+                Danh sách ngẫu nhiên được tạo khi tải trang để minh họa giao diện.
+              </Paragraph>
+              <Row gutter={[16, 16]}>
+                {products.map((p) => (
+                  <Col key={p.id} xs={24} sm={12} md={8} lg={6}>
+                    <Card hoverable className="!rounded-xl h-full" title={p.name}>
+                      <Paragraph className="!mb-4 text-gray-600">{p.desc}</Paragraph>
+                      <div className="text-xl font-semibold text-blue-600">
+                        {p.price.toLocaleString("vi-VN")}₫
+                      </div>
+                      <div className="mt-4">
+                        <Button type="primary" block className="!rounded-lg">Dùng thử</Button>
+                      </div>
+                    </Card>
+                  </Col>
+                ))}
+              </Row>
             </div>
           </div>
-        </PageContainer>
-      </ProLayout>
-    </div>
+        </div>
+      </PageContainer>
+    </Header>
   );
 };
 export default HomePage;
